+++ CHANGES.md
@@ -7,7 +7,10 @@
 
   * SBV now supports cvc5; the latest incarnation of CVC. See https://github.com/cvc5/cvc5
     for details.
-    
+
+  * SBV now supports bitwuzla; the latest incarnation of Boolector. See https://github.com/bitwuzla/bitwuzla
+    for details.
+
   * Fixed handling of CRational values in constant folding, which was missing a case.
     Thanks to Jaro Reinders for reporting.